changed to float

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-income = int(input("Enter your monthly income:"))
-expense = int(input('Enter your total monthly expenses:'))
+income = float(input("Enter your monthly income:"))
+expense = float(input('Enter your total monthly expenses:'))
 monthly_savings = income - expense
 projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
 print("Your monthly savings are $" + str(monthly_savings))
